[WEB-1450] - Moved files out of "shared" folder (https://github.com/thegoodparty/gp-webapp/pull/122#discussion_r1532361008)
