+++ app/admin/top-issues/components/DynamicSVGIcon.js
